+++ lib/camera/camera.dart
@@ -1,7 +1,6 @@
 import 'dart:math' as math;
 import 'dart:ui';
 
-import 'package:bonfire/base/bonfire_game.dart';
 import 'package:bonfire/base/game_component.dart';
 import 'package:bonfire/bonfire.dart';
 import 'package:bonfire/util/bonfire_game_ref.dart';
@@ -10,7 +9,7 @@ import 'package:flutter/widgets.dart';
 
 import 'camera_config.dart';
 
-class Camera with BonfireHasGameRef<BonfireGame> {
+class Camera with BonfireHasGameRef {
   bool _isMoving = false;
   double _spacingMap = 32.0;
   Offset position = Offset.zero;