+++ lib/collision/object_collision.dart
@@ -71,10 +71,11 @@ mixin ObjectCollision on GameComponent {
   @override
   void render(Canvas canvas) {
     super.render(canvas);
-    if (gameRef.showCollisionArea == true) {
+    if ((gameRef as BonfireGame).showCollisionArea == true) {
       _drawCollision(
         canvas,
-        gameRef.collisionAreaColor ?? Colors.lightGreen.withOpacity(0.5),
+        (gameRef as BonfireGame).collisionAreaColor ??
+            Colors.lightGreen.withOpacity(0.5),
       );
     }
   }