+++ lib/game_interface/game_interface.dart
@@ -56,8 +56,8 @@ class GameInterface extends GameComponent {
   }
 
   void _drawFPS(Canvas c) {
-    if (gameRef.showFPS == true) {
-      double? fps = gameRef.fps(100);
+    if ((gameRef as BonfireGame).showFPS == true) {
+      double? fps = (gameRef as BonfireGame).fps(100);
       getTextConfigFps(fps).render(
         c,
         'FPS: ${fps.toStringAsFixed(2)}',