+++ lib/util/direction_animations/simple_direction_animation.dart
@@ -1,3 +1,4 @@
+import 'dart:async';
 import 'dart:ui';
 
 import 'package:bonfire/bonfire.dart';
@@ -9,10 +10,10 @@ import 'package:flutter/foundation.dart';
 
 /// Class responsible to manager animation on `SimplePlayer` and `SimpleEnemy`
 class SimpleDirectionAnimation {
-  late SpriteAnimation idleLeft;
-  late SpriteAnimation idleRight;
-  late SpriteAnimation runLeft;
-  late SpriteAnimation runRight;
+  SpriteAnimation? idleLeft;
+  SpriteAnimation? idleRight;
+  SpriteAnimation? runLeft;
+  SpriteAnimation? runRight;
 
   SpriteAnimation? idleUp;
   SpriteAnimation? idleDown;
@@ -41,23 +42,23 @@ class SimpleDirectionAnimation {
   double opacity = 1.0;
 
   SimpleDirectionAnimation({
-    required Future<SpriteAnimation> idleLeft,
-    required Future<SpriteAnimation> idleRight,
-    required Future<SpriteAnimation> runRight,
-    required Future<SpriteAnimation> runLeft,
-    Future<SpriteAnimation>? idleUp,
-    Future<SpriteAnimation>? idleDown,
-    Future<SpriteAnimation>? idleUpLeft,
-    Future<SpriteAnimation>? idleUpRight,
-    Future<SpriteAnimation>? idleDownLeft,
-    Future<SpriteAnimation>? idleDownRight,
-    Future<SpriteAnimation>? runUp,
-    Future<SpriteAnimation>? runDown,
-    Future<SpriteAnimation>? runUpLeft,
-    Future<SpriteAnimation>? runUpRight,
-    Future<SpriteAnimation>? runDownLeft,
-    Future<SpriteAnimation>? runDownRight,
-    Map<String, Future<SpriteAnimation>>? others,
+    required FutureOr<SpriteAnimation> idleLeft,
+    required FutureOr<SpriteAnimation> idleRight,
+    required FutureOr<SpriteAnimation> runRight,
+    required FutureOr<SpriteAnimation> runLeft,
+    FutureOr<SpriteAnimation>? idleUp,
+    FutureOr<SpriteAnimation>? idleDown,
+    FutureOr<SpriteAnimation>? idleUpLeft,
+    FutureOr<SpriteAnimation>? idleUpRight,
+    FutureOr<SpriteAnimation>? idleDownLeft,
+    FutureOr<SpriteAnimation>? idleDownRight,
+    FutureOr<SpriteAnimation>? runUp,
+    FutureOr<SpriteAnimation>? runDown,
+    FutureOr<SpriteAnimation>? runUpLeft,
+    FutureOr<SpriteAnimation>? runUpRight,
+    FutureOr<SpriteAnimation>? runDownLeft,
+    FutureOr<SpriteAnimation>? runDownRight,
+    Map<String, FutureOr<SpriteAnimation>>? others,
     SimpleAnimationEnum initAnimation = SimpleAnimationEnum.idleRight,
   }) {
     _currentType = initAnimation;